refactor: move audio source selection out of the ViewModel
Adds AudioSourceResolver, which owns the third-party chain and picks the source,
so the four resolve sites in PlaybackViewModel each become one call and the
ViewModel no longer branches on the audio-source setting.

Trims comments to the facts that are not recoverable from the code, and drops
the login-required fixture in favour of an inline literal.
