--- src/app/src/test/resources/fixtures/youtubemusic/player_login_required.json
@@ -1,6 +1,0 @@
-{
- "playabilityStatus": {
-  "status": "LOGIN_REQUIRED",
-  "reason": "Sign in to confirm you're not a bot"
- }
-}
